+++ interpreter/closure.hpp
@@ -48,7 +48,7 @@ namespace interpreter {
         // CallableValue:
         virtual void call(Process& proc, afl::data::Segment& args, bool want_result);
         virtual bool isProcedureCall() const;
-        virtual int32_t getDimension(int32_t which) const;
+        virtual size_t getDimension(size_t which) const;
         virtual Context* makeFirstContext();
 
         // BaseValue:
